requirements.txt
asgiref==3.2.7
astroid==2.4.2
colorama==0.4.3
dj-database-url==0.5.0
Django==2.1.15
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-userforeignkey==0.3.0
djangorestframework==3.11.1
gunicorn==20.0.4
html5lib==1.1
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.6.0
PyPDF2==1.26.0
python-decouple==3.3
pytz==2020.1
reportlab==3.5.49
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
Unipath==1.1
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.4
